Update insights-custom-repos.yml
I suggest adding this note because the "Build an image with custom content" quick start Insights Learning Resources page (https://console.redhat.com/insights/learning-resources) is valid only in the Preview mode. The first procedure "Add a Custom Repository" in the  Quick start doesn't work today unless preview mode is enabled. This is a temporary situation. When the HMS feature is GA, this will work as documented.
Because the quick start framework doesn't provide a way to show a Quick Start card only for the preview mode, I suggest also adding a step that tells the user to enable preview mode before proceeding with the rest of the quick start procedures. 

I have set a note to remember to remove the preview step when the features described in the Quick Start are GA.